% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortunator.R
\name{fortunator}
\alias{fortunator}
\title{Create a fortune}
\usage{
fortunator(seed = NULL)
}
\arguments{
\item{seed}{optional seed.}
}
\description{
Creates a random fortune-like quote based on word frequency derived from the dataset in the
fortunes package.
}
\examples{

fortunator(-814)
fortunator(943)
fortunator(-499)
fortunator(c(339, 12450))

}
